Criei a tela SplashFragment
