+++ app/src/main/java/com/example/loginmvpmvvm/ui/home/HomeFragment.kt
@@ -1,14 +1,12 @@
 package com.example.loginmvpmvvm.ui.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
-import com.example.loginmvpmvvm.R
+import androidx.fragment.app.Fragment
 import com.example.loginmvpmvvm.databinding.FragmentHomeBinding
-import com.example.loginmvpmvvm.databinding.FragmentLoginBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class HomeFragment : Fragment() {
@@ -17,12 +15,12 @@ class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private val binding get() = _binding!!
 
-//    override fun onCreate(savedInstanceState: Bundle?) {
-//        super.onCreate(savedInstanceState)
-//        activity?.intent?.getStringExtra("id")?.let {
-//            viewModel.getUser(it)
-//        }
-//    }
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        activity?.intent?.getStringExtra("id")?.let {
+            viewModel.getUser(it)
+        }
+    }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
