+++ app/src/main/res/navigation/nav_app.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_app"
+    app:startDestination="@id/nav_splash_fragment">
+
+    <fragment
+        android:id="@+id/nav_splash_fragment"
+        android:name="com.example.loginmvpmvvm.ui.splash.SplashFragment"
+        tools:layout="@layout/fragment_splash">
+        <action
+            android:id="@+id/action_splashFragment_to_nav_access_fragment"
+            app:destination="@id/nav_access_fragment"
+            app:enterAnim="@android:anim/slide_in_left"
+            app:exitAnim="@android:anim/slide_out_right" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/nav_access_fragment"
+        android:name="com.example.loginmvpmvvm.ui.access.AccessFragment"
+        android:label="fragment_access"
+        tools:layout="@layout/fragment_access" >
+        <action
+            android:id="@+id/action_accessFragment_to_loginFragment"
+            app:destination="@id/nav_login_fragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/nav_login_fragment"
+        android:name="com.example.loginmvpmvvm.ui.access.login.LoginFragment"
+        android:label="fragment_login"
+        tools:layout="@layout/fragment_login">
+        <action
+            android:id="@+id/action_loginFragment_to_signUpFragment"
+            app:destination="@id/nav_sign_up_fragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/nav_sign_up_fragment"
+        android:name="com.example.loginmvpmvvm.ui.access.signup.SignUpFragment"
+        android:label="SignUpFragment"
+        tools:layout="@layout/fragment_signup">
+        <action
+            android:id="@+id/action_signUpFragment_to_loginFragment"
+            app:destination="@id/nav_login_fragment" />
+    </fragment>
+</navigation>
